requirements.txt
requests>=2.32.2
pandas>=2.2.2
python-dotenv>=1.0.1
tenacity>=8.3.0
rake-nltk>=1.0.6
nltk>=3.9.1
dateparser>=1.2.0
tqdm>=4.66.4
scikit-learn